worker: cleanup and reorganize
